requirements.txt
absl-py==2.0.0
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
ase==3.22.1
asttokens==2.4.1
async-lru==2.0.4
atom3d==0.2.6
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
biopython==1.81
bleach==6.1.0
blosc2==2.0.0
cachetools==5.3.2
certifi
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
comm==0.2.0
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dgl==1.1.3
dill==0.3.7
docker-pycreds==0.4.0
docopt==0.6.2
easy-parallel==0.1.6
easydict==1.11
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.45.1
fqdn==1.5.1
freesasa==2.2.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.0.0
grpcio==1.59.3
h5py==3.10.0
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
lightning-utilities==0.10.0
lit==17.0.6
lmdb
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
ml-collections==0.1.1
mpmath==1.3.0
msgpack==1.0.7
multipledispatch==1.0.0
multiprocess==0.70.15
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.6
numpy==1.20.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
pox==0.3.3
ppft==1.7.6.7
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
pyrr==0.10.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rdkit==2023.9.3
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
Send2Trash==1.8.2
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tables==3.8.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
torch==1.13.1
torch-tb-profiler==0.4.3
torch_geometric==2.4.0
torchmetrics==1.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
# unicore==0.0.1
uri-template==1.3.0
urllib3==2.1.0
wandb==0.16.1
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yarg==0.1.9
zipp==3.17.0
